+++ src/main/kotlin/com/lambda/interaction/material/container/containers/ChestContainer.kt
@@ -23,9 +23,9 @@ import com.lambda.interaction.handlers.ContainerHandler
 import com.lambda.interaction.material.container.ExternalContainer
 import com.lambda.interaction.material.container.MaterialContainer
 import com.lambda.task.Task
-import com.lambda.task.TaskSupplier
-import com.lambda.task.thenAction
+import com.lambda.task.wrappers.TaskSupplier
 import com.lambda.task.tasks.OpenContainerTask
+import com.lambda.task.wrappers.thenAction
 import com.lambda.util.extension.containerSlots
 import com.lambda.util.text.buildText
 import com.lambda.util.text.highlighted
