+++ src/main/kotlin/com/lambda/module/modules/player/AutoEat.kt
@@ -28,7 +28,7 @@ import com.lambda.module.tag.ModuleTag
 import com.lambda.task.RootTask.run
 import com.lambda.task.Task
 import com.lambda.task.tasks.EatTask.Companion.eat
-import com.lambda.task.thenAction
+import com.lambda.task.wrappers.thenAction
 import com.lambda.threading.runSafeAutomated
 
 @Suppress("unused")
