+++ src/main/kotlin/com/lambda/module/modules/world/AutoVillagerCycle.kt
@@ -36,8 +36,8 @@ import com.lambda.module.tag.ModuleTag
 import com.lambda.sound.SoundHandler.playSound
 import com.lambda.task.RootTask.run
 import com.lambda.task.Task
-import com.lambda.task.thenAction
 import com.lambda.task.tasks.BuildTask.Companion.build
+import com.lambda.task.wrappers.thenAction
 import com.lambda.threading.runSafeAutomated
 import com.lambda.util.BlockUtils.blockState
 import com.lambda.util.BlockUtils.isEmpty
@@ -280,9 +280,7 @@ object AutoVillagerCycle : Module(
 			buildTask = runSafeAutomated {
 				lecternPos.toStructure(TargetState.Empty)
 					.build(finishOnDone = true)
-					.thenAction {
-						switchState(CycleState.PlaceLectern)
-					}
+					.thenAction { switchState(CycleState.PlaceLectern) }
 					.run()
 			}
 			switchState(CycleState.WaitBreak)
