+++ src/main/kotlin/com/lambda/task/tasks/ContainerTransferTask.kt
@@ -28,7 +28,7 @@ import com.lambda.interaction.material.container.ExternalContainer
 import com.lambda.interaction.material.container.MaterialContainer
 import com.lambda.interaction.material.container.containers.InventoryContainer
 import com.lambda.task.Task
-import com.lambda.task.thenAction
+import com.lambda.task.wrappers.thenAction
 import com.lambda.threading.runSafeAutomated
 
 class ContainerTransferTask(
@@ -54,9 +54,9 @@ class ContainerTransferTask(
 				val toSlots = toContainer.slots
 				fromExternal.takeIf { slots.isEmpty() }?.let { fromExternal ->
 					if (toContainer is ExternalContainer && toSlots.isEmpty()) {
-						fromContainer.transferByTask(stackSelection, InventoryContainer).thenAction {
-							fromContainer = InventoryContainer
-						}.execute(this@ContainerTransferTask)
+						fromContainer.transferByTask(stackSelection, InventoryContainer)
+							.thenAction { fromContainer = InventoryContainer }
+							.execute(this@ContainerTransferTask)
 						return@listen
 					}
 					delegateTask = fromExternal.accessThen {
